api/requirements.txt
bcrypt==3.2.2
numpy==1.18.3
opensearch-py==1.0.0
passlib==1.7.4
PyJWT==2.4.0
python-dateutil==2.8.2
requests==2.23.0
requests_aws4auth==0.9
